Count groups

A helper function to count groups to add to random effect results.

count_grps(model, grp_vars)

# S3 method for default
count_grps(model, grp_vars)

# S3 method for merMod
count_grps(model, grp_vars)

# S3 method for glmmTMB
count_grps(model, grp_vars)

# S3 method for lme
count_grps(model, grp_vars)

# S3 method for brmsfit
count_grps(model, grp_vars)

# S3 method for stanreg
count_grps(model, grp_vars)

# S3 method for stanmvreg
count_grps(model, grp_vars)

# S3 method for gam
count_grps(model, grp_vars)

Arguments

model

A fitted model e.g. from lme4.

grp_vars

A character vector for the grouping/cluster variables used for random effects

Value

A tibble of the results.

Details

For each grouping variable for which random effects are estimated, count the respective group sizes. This is not meant to be used directly.

Examples

library(lme4)
#> Loading required package: Matrix
library(mixedup)

mod = lmer(Reaction ~ Days + (1 + Days | Subject), sleepstudy)

count_grps(mod, 'Subject')
#> # A tibble: 18 × 3
#>    group_var group     n
#>    <chr>     <chr> <int>
#>  1 Subject   308      10
#>  2 Subject   309      10
#>  3 Subject   310      10
#>  4 Subject   330      10
#>  5 Subject   331      10
#>  6 Subject   332      10
#>  7 Subject   333      10
#>  8 Subject   334      10
#>  9 Subject   335      10
#> 10 Subject   337      10
#> 11 Subject   349      10
#> 12 Subject   350      10
#> 13 Subject   351      10
#> 14 Subject   352      10
#> 15 Subject   369      10
#> 16 Subject   370      10
#> 17 Subject   371      10
#> 18 Subject   372      10