extract_nlme_variances

Because nlme won't do it for you.

extract_het_var(model, ...)

Arguments

model
An appropriate mixed model

...
Other arguments appropriate to the method

Value

A vector of the estimates on the variance scale.

Details

For models with heterogeneous variance, i.e. that contain something like varIdent(form = ~1|Group), nlme returns a result in the summary regarding the variances that many do not know what to do with, nor likely would be what they would want to report. I rarely do these models with nlme, and have only played around with the varIdent case. This function ' will save me a little trouble in demos I do, but I have no plans to do any more with it.

Examples

library(nlme)

#> 
#> Attaching package: 'nlme'

#> The following object is masked from 'package:lme4':
#> 
#>     lmList

library(mixedup)

model <- lme(distance ~ age + Sex, data = Orthodont, random = ~ 1|Subject,
weights=varIdent(form = ~1|Sex))

summary(model)

#> Linear mixed-effects model fit by REML
#>  Data: Orthodont 
#>        AIC      BIC    logLik
#>   432.3567 448.2805 -210.1784
#> 
#> Random effects:
#>  Formula: ~1 | Subject
#>         (Intercept) Residual
#> StdDev:    1.839463 1.761758
#> 
#> Variance function:
#>  Structure: Different standard deviations per stratum
#>  Formula: ~1 | Sex 
#>  Parameter estimates:
#>      Male    Female 
#> 1.0000000 0.4541324 
#> Fixed effects: distance ~ age + Sex 
#>                 Value Std.Error DF   t-value p-value
#> (Intercept) 18.919992 0.7285568 80 25.969138  0.0000
#> age          0.549887 0.0473096 80 11.623168  0.0000
#> SexFemale   -2.321023 0.7629703 25 -3.042088  0.0055
#>  Correlation: 
#>           (Intr) age   
#> age       -0.714       
#> SexFemale -0.468  0.000
#> 
#> Standardized Within-Group Residuals:
#>         Min          Q1         Med          Q3         Max 
#> -3.25494115 -0.53013324 -0.02554914  0.51114273  3.03915818 
#> 
#> Number of Observations: 108
#> Number of Groups: 27 

extract_het_var(model)

#>      Male    Female
#> 1 3.10379 0.6401141