Extract variance components

This has functionality for simpler models from lme4, glmmTMB, nlme, and brms.

extract_vc(model, ci_level = 0.95, ci_args = NULL, ci_scale = "sd",
  show_cor = FALSE, digits = 3, component = "cond", ...)

# S3 method for merMod
extract_vc(model, ci_level = 0.95, ci_args = NULL,
  ci_scale = "sd", show_cor = FALSE, digits = 3, ...)

# S3 method for glmmTMB
extract_vc(model, ci_level = 0.95, ci_args = NULL,
  ci_scale = "sd", show_cor = FALSE, digits = 3,
  component = "cond", ...)

# S3 method for lme
extract_vc(model, ci_level = 0.95, ci_args = NULL,
  ci_scale = "sd", show_cor = FALSE, digits = 3, ...)

# S3 method for brmsfit
extract_vc(model, ci_level = 0.95, ci_args = NULL,
  ci_scale = "sd", show_cor = FALSE, digits = 3, ...)

# S3 method for gam
extract_vc(model, ci_level = 0.95, ci_args = NULL,
  ci_scale = "sd", show_cor = FALSE, digits = 3, ...)

Arguments

model
an lme4, glmmTMB, nlme, mgcv, or brms model

ci_level
confidence level < 1, typically above 0.90. A value of 0 will not report it (except for gam objects, which will revert to .95 due to gam.vcomp). Default is .95.

ci_args
Additional arguments to the corresponding confint method.

ci_scale
A character string of 'sd' or 'var' to note the scale of the interval estimate. Default is 'sd'. at present.

show_cor
Return the intercept/slope correlations as a separate list element. Default is FALSE.

digits
Rounding. Default is 3.

component
For glmmTMB objects, which of the three components 'cond', 'zi' or 'other' to select. Default is cond. Minimal testing on other options.

...
Other stuff to pass to the corresponding function. Unused/tested

Value

A data frame with output for variance components or list that also contains the correlations of the random effects.

Details

Returns a more usable (my opinion) version of variance components estimates including variance, standard deviation, the confidence interval for either, the relative proportion of variance, and all in a data frame with names that are clean and easy to use.

See also

Examples

library(lme4)
library(mixedup)

lmer_mod <- lmer(Reaction ~ Days + (1 + Days | Subject), data = sleepstudy)

extract_vc(lmer_mod)

#> Computing profile confidence intervals ...

#>      group    effect variance     sd sd_2.5 sd_97.5 var_prop
#> 1  Subject Intercept  611.898 24.737 14.382  37.716    0.470
#> 2  Subject      Days   35.081  5.923  3.801   8.753    0.027
#> 3 Residual            654.941 25.592 22.898  28.858    0.503

extract_vc(lmer_mod, ci_scale = 'var')

#> Computing profile confidence intervals ...

#>      group    effect variance     sd var_2.5 var_97.5 var_prop
#> 1  Subject Intercept  611.898 24.737 206.833 1422.496    0.470
#> 2  Subject      Days   35.081  5.923  14.449   76.621    0.027
#> 3 Residual            654.941 25.592 524.331  832.784    0.503